Modified to use bower directly.
Using Programmatic API of bower to control 'offline' mode.
I want to handle bower dependencies like Gradle's
- once cached, it's installable without network
connection - but bower's offline mode works globally,
so added a small script to control that.
From this change, dependencies are installed sequentially
and if it's already cached, install with offline mode.
Otherwise (not cached), it will be installed with network
connection as usual.