Improve error handling for bower (#3)
process.exit(1) does not work as expected,
so I changed it to listening exit event and call exit from
the listener.
In error handlers for bower commands,
instead of calling process.exit(1), I use deferred.reject().
Although this fix is applied, there is still an issue
that bower does not report download errors as expected
in some cases.