Upgrade bower to v1.7.6 and downgrade node to 0.10
bower version 1.6.3 that is included in the plugin
often fails to install dependencies without raising errors.
I also confirmed this kind of behavior on bower 1.6.3
running on Node.js.
bower 1.7.6 handles this better (though I'm not sure that it's the best),
so I updated bower to 1.7.6.
Trireme 0.8.8 with node version 0.12 cannot run bower 1.7.6
because of node 'which' module's git availability check,
and node 0.10 works well with that.
Usually downgrading dependency is not a good choice,
but it's better than now.