Fix that parallel cache checking failure
Currently, transitive dependencies' versions are
checked only with the top level dependencies.
For example, if we define bootstrap that
has jquery dependency, we must also define jquery
in bower.dependencies. Otherwise, jquery is always
assumed to be not cached and installation will be
always online.
In the future, transitive dependencies should be
checked with bower.commands.cache() as top level
dependencies.